requirements.txt
--- Core Data & Numerics ---

pandas>=2.0.0
numpy>=1.24.0

--- Scraping & Network (Inherited from nba_scraper) ---

requests>=2.31.0
urllib3>=2.0.0
PyYAML>=6.0

--- Analytics & Storage (Inherited from nba_parser) ---

scikit-learn is used for Ridge Regression in RAPM calculations

scikit-learn>=1.3.0

SQLAlchemy is used if you plan to dump stats to a local DB

SQLAlchemy>=2.0.0

--- Development & Testing ---

Essential for the Agent to run tests and format code

pytest>=7.4.0
black>=23.0.0
ruff>=0.1.0

--- Typing ---

Helps the Agent understand external library types

types-requests
types-PyYAML